float for normalization coefficients

+++ cm_fit/data_loader/data_utils.py
@@ -41,5 +41,5 @@ def string_to_list(input_line):
     output_line = input_line.split(" [")[-1]
     output_line = output_line.replace("]", "")
     output_line = output_line.split(", ")
-    output_line = [int(item) for item in output_line]
+    output_line = [float(item) for item in output_line]
     return output_line
